refactor(engine): extract refresh coordinator test fixtures into sibling module
Three repo-wrapper fixtures (FlakyReleaseRepo, AlwaysContendedRepo,
AlwaysFailHeartbeatRepo) each forwarded 6-7 RefreshClaimRepo trait
methods to an inner repo, totalling ~165 lines of forwarding boilerplate
inside `coordinator.rs::tests`. This was the file-size growth flagged by
the wave-3 code-quality reviewer.

Move them to `crates/engine/src/credential/refresh/test_fixtures.rs`
gated on `#[cfg(test)]`. The coordinator's test module imports the three
types via a single `use crate::credential::refresh::test_fixtures::{..}`.

coordinator.rs: 1583 -> 1417 lines (-166).
test_fixtures.rs: new, 197 lines.

All 20 coordinator tests + 277 nebula-engine tests still pass.
